LayoutTests/fast/js/toString-for-var-decl-expected.txt

This test checks that toString() round-trip does not change meaning of functions that contain var declarations inside for -loop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ueuf1 is uf1
PASS ueuf2 is uf2
PASS uf2 != uf3 is true
PASS successfullyParsed is true

TEST COMPLETE
